requirements.txt
langchain
langchain-core
langchain-huggingface
transformers
huggingface-hub
python-dotenv
numpy
scikit-learn
Torch
streamlit
fastapi
uvicorn
pymongo
datetime